statement_generator/requirements.txt
anthropic>=0.40.0
pydantic>=2.0.0
click>=8.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
black>=23.0.0
mypy>=1.0.0
